Coll № 1/Col 1 Task 3/src/LinkedListReversal.java
class Node {
    int data;
    Node next;

    public Node(int data) {
        this.data = data;
        this.next = null;
    }
}

@SuppressWarnings("ClassEscapesDefinedScope")
public class LinkedListReversal {

    public static Node reverseLinkedList(Node head) {
        Node prev = null;
        Node current = head;
        Node next;

        while (current != null) {
            next = current.next;
            current.next = prev;
            prev = current;
            current = next;
        }

        head = prev;
        return head;
    }

    public static void printLinkedList(Node head) {
        Node current = head;

        while (current != null) {
            System.out.print(current.data + " ");
            current = current.next;
        }
        System.out.println();
    }

    public static void main(String[] args) {
        Node head = new Node(1);
        head.next = new Node(2);
        head.next.next = new Node(3);
        head.next.next.next = new Node(4);

        System.out.println("Исходный связный список:");
        printLinkedList(head);

        head = reverseLinkedList(head);

        System.out.println("Развернутый связный список:");
        printLinkedList(head);
    }
}
